task11

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -12,7 +12,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import jp.co.seattle.library.dto.RentBookDetailsInfo;
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentService;
 
 /**
  * 削除コントローラー
@@ -24,6 +26,9 @@ public class DeleteBookController {
 
 	@Autowired
 	private BooksService booksService;
+
+	@Autowired
+	private RentService rentService;
 
 	/**
 	 * 対象書籍を削除する
@@ -38,11 +43,19 @@ public class DeleteBookController {
 	public String deleteBook(Locale locale, @RequestParam("bookId") Integer bookId, Model model) {
 		logger.info("Welcome delete! The client locale is {}.", locale);
 
-		booksService.deleteBook(bookId);
+		RentBookDetailsInfo selectedRentBookInfo = rentService.selectRentBookInfo(bookId);
 
-		model.addAttribute("bookList", booksService.getBookList());
-		return "home";
+		if (!(selectedRentBookInfo == null)) {
+			model.addAttribute("errorMessage", "貸し出し済みです");
+			model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+			return "details";
+
+		} else {
+			booksService.deleteBook(bookId);
+			model.addAttribute("bookList", booksService.getBookList());
+			return "home";
+
+		}
 
 	}
-
 }
